Add files via upload
Calbia #1
This is a fun-learning project about a fictional country, located in the South of the Cheetos continent, created by me and my team. We have worked upon its History, Legislature, Executive System, Judiciary, and Foreign Policy. For the project, we borrowed ideas from the constitutions of Kenya, Germany, India, Australia, the United States of America, the United Kingdom, Ghana, Newzealand, Spain, France, and South Africa. Not to forget, the Historical inspiration is heavily influenced by The War of Roses, Spanish history, Early Egyptian History, Mongolian History, and German History. We've had a lot of fun in coming up with our own History which happens in a parallel world, by the way. We've gone full GRRM on the History. Okay, probably not as good as him! Our idea was to come up with a country with the least number of loopholes in the Legislature, Executive, and Judicial system.